+++ CHANGES.rst
@@ -5,6 +5,9 @@
 Unreleased
 ==========
 
+- Switch to using non-backward-compatible pickles (protocol 3, without
+  storing bytes as strings) under Python 3.
+
 - Fixed: A ``UnicodeDecodeError`` could happen for non-ASCII OIDs
   when using bushy blob layout.